provider: client handling - legacy_proposals
