+++ x/ccv/provider/client/cli/tx.go
@@ -49,6 +49,7 @@ func NewAssignConsumerKeyCmd() *cobra.Command {
 				return err
 			}
 
+			signer := clientCtx.GetFromAddress().String()
 			txf, err := tx.NewFactoryCLI(clientCtx, cmd.Flags())
 			if err != nil {
 				return err
@@ -57,7 +58,7 @@ func NewAssignConsumerKeyCmd() *cobra.Command {
 
 			providerValAddr := clientCtx.GetFromAddress()
 
-			msg, err := types.NewMsgAssignConsumerKey(args[0], sdk.ValAddress(providerValAddr), args[1])
+			msg, err := types.NewMsgAssignConsumerKey(args[0], sdk.ValAddress(providerValAddr), args[1], signer)
 			if err != nil {
 				return err
 			}
